Added G_GlobalSound. Takes one argument, the soundIndex of the sound that is to be played.
I do not recommend nesting G_SoundIndex inside of the function when calling it.  Instead G_SoundIndex should be used to return an integer, then call G_GlobalSound with that integer.

I added a return to avoid any S_FindName errors that can occur if it's called without a proper index. 

I have not committed the declaration to g_local since this is actually part of a bigger change--I am adding server configurable killing sprees. See the wiki pages later on this week for more information.