Multikill sounds are now disabled/enabled on the client by having the cvar g_altExcellent flagged as CVAR_SERVERINFO. All checks that would revert multikills to stock behavior now function by using the trap_Cvar_Set function. Added the check to disable multikills if its set two 1 and we didn't read any multikills. NO MORE CONFIG STRINGS!!!! :)
The g_spreeDiv cvar now dumps its value to level.spreeDivisor when sprees and kills are read.  The check for "0" or "1" still happens, but level.spreeDivisor is set to the default of 5 instead of the cvar.  The cvar is left alone, and if its not corrected by the server owner, it will just keep printing a "revert to default" type message when reading sprees/multikills. This removes any possibility of a server owner creating a divide by 0 situation. 

Cleaned up a couple of things on some functions as well, mostly redundant logic bits.