Added parsing of a maps/<mapname>.info (all lowercase). This allows arena like information but can easily be accessed from UI and game-logic. It allows the gamelogic and UI to share information about bot setup and things like that.
And example file for oasage2 could be:
auther Sago007
captureLimit 8
minTeamSize 2
gametype 1fctf
minTeamSize 3

note that the gametype line causes the folling commands to only be used for a given gametype. If * is used all gametypes match if an unknown gametype is given nothing matches this allows furture arguments to be included in the file. 
File size is currently limited to 4 KiB